Refactor: ClipStartPage file structure moved, App.jsx import refactor
